requirements.txt
Flask==3.0.*
line-bot-sdk==3.0.*
requests==2.31.*
gunicorn==22.0.*
